asset-registry: And and Implement `Inspect` + `Mutate` traits (#767)
* traits: add asset_registry::{Inspect, Mutate}

* Move to use traits::asset_registry::AssetMetadata

The AssetMetadata type is needed by the Inspect and Mutate traits so it
needs to be in a shared place. While it's fine for the pallet to depend
on the traits::asset_registry, the inverse not so much, so we move the
struct to the traits section and import it here.

* asset-registry: Implement Inspect & Mutate

* fixup: use sp_std::vec::Vec;

* review: clean up imports

* asset-registry: pub use AssetMetadata

Doing so makes logical sense and smoothes the update to this new
version.